no free list for address sanitizer
doesn't realize that the memory is not
stack memory anymore after fiber finished

Didn't help:
When leaving a fiber definitely, null must be passed as first
argument to the start_switch function so that the fake stack is destroyed.
If you do not want support for stack use-after-return detection, you can
always pass null to these two functions.